Add webpack, new file structure and test cases
