+++ src/js/GameBoard.js
@@ -1,0 +1,24 @@
+const gameBoard = (() => {
+  const board = [' ', ' ', ' ', ' ', ' ', ' ', ' ', ' ', ' '];
+
+  const getElement = (index) => board[index];
+
+  const setElement = (data, index) => {
+    board[index] = data;
+  };
+
+  const resetBoard = () => {
+    board.forEach((element, index) => {
+      setElement(' ', index);
+    });
+  };
+
+  return {
+    getElement,
+    setElement,
+    resetBoard,
+    board,
+  };
+})();
+
+export default gameBoard;
